feat: Added the context for the DropDownMaid box.
When you go from /maid to /maid/pickup through services from the navbar the context is being preserved.

When you go from /maid to /maid/pickup through the button "Get maid now" from the DropDownBoxMaid the context is not preserved.

I think it is due to the fact when you go from the DropDownBoxMaid it re-renders the left side of the page.